requirements.txt
# Google APIs
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
gspread==5.12.3

# OCR Libraries
pytesseract==0.3.10
easyocr==1.7.1

# Image Processing
opencv-python-headless==4.8.1.78
Pillow==10.1.0
numpy==1.24.3

# Data Processing
pandas==2.1.4

# Utilities
python-dateutil==2.8.2
